switch from big endian to little endian in the encoded header
It almost certainly doesn't really matter which one we pick. I
originally went with big endian because that's the vaguely standardized
"network byte order". I'm switching to little endian because the BLAKE2
standard made the same switch relative to BLAKE
(https://blake2.net/blake2.pdf), and that seems like a very relevant
precedent to follow. Little endian is also what DJB chose for his NaCl
library.

The main downside of little endian in my mind, is that because it's also
the most common native endianness in practice, implementations are more
likely to write code that "happens to work" on their machine and fails
on big endian architectures. In fact this totally happened in the
official implementation of BLAKE2 (https://github.com/BLAKE2/libb2/issues/12).
However, a contributing factor in that case was the bug only affected a
couple of obscure options, which might actually never have been
exercised on any big endian machines, ever. This isn't the case for bao;
if you flip the endianness of your header length, then almost all of
your hashes will be wrong.