stop trying to flip the encoded file on Drop
That's a huge expensive operation, and it's inappropriate to be doing it
in a destructor. The obvious case for dropping the encoder without
flipping is if one of the initial writes failed, in which case we
shouldn't be doing any more work anyway.

Also start cleaning up the unverified.rs file. Eventually we'll get rid
of that.