get rid of the main function in test_bao.py
It was redundant with pytest, and it ran the risk of getting out of sync
someday. In theory it might've made it easier to run Python tests on CI,
without having to deal with installing pytest. But Travis being stuck on
Python 3.4 by default (blake2b came in 3.6) is such a drag anyway, that
for now we're not going to run Python on CI at all.