stop converting stdin/stdout to File
Previously it was *very* convenient to treat all forms of IO as File.
(Though that did mean main.rs was littered with confirm_real_file
calls.) But that all rested on an os_pipe API decision that turned out
to be not totally sound.

As noted in the os_pipe 0.8 change notes, the standard library's
handling of pipes-as-files on Windows isn't very graceful. Read and
write can work, but metadata() doesn't always work (for pipes?), and
when it does (for ttys?) it can lead to is_file() incorrectly returning
true.

So rather than deal with all that, this commit takes the high road and
avoids confusing pipes and files to begin with. That requires a little
more bookkeeping in main.rs, but I think the result is less error prone,
because calling .required_file() is mandatory to get a Seek impl.