add some more color to animation.py and rerecord bao_hash.cast
Note that to get the performance numbers in that recording, I do a
`sudo systemctl disable gdm` to turn off the GUI entirely and then
restart my laptop. With the desktop running, performance is about 10%
worse (for everything, not just for bao). Capturing the result with
asciinema also lets me replay it in a GUI terminal with a screen
recorder, without the screen recording itself hurting hash
performance.