change the max depth parameter from 64 to 255
Previously we used 64 to indicate that the max input length was 2^64 - 1
bytes. (Though taking into account the actual chunk size, the true
maximum depth was 52.) However, that wasn't the right interpretation of
the maximum depth parameter according to the BLAKE2 spec. In the spec,
inputs that saturate the maximum depth of the tree are supposed to
increase the fanout of the root node. Bao will never do anything like
that. A Bao implementation is within its rights to panic given an input
longer than 2^64 - 1 bytes -- which is already so astronomically large
that it could never be hashed in practice -- but if it does produce an
answer it's expected to keep making the tree deeper rather than
increasing fanout. In the BLAKE2 spec, this "unlimited depth" behavior
is denoted by setting the max depth parameter to 255.