get rid of the all-at-once encode implementations
Replace them with the encode::encode() and encode::outboard()
convenience wrappers around Writer.

The all-at-once functions were a maintenance and testing burden, for
arguably very little benefit. As I get ready to experiment with
node_offset, it's important to trim this stuff down. I'm about to do the
same for decode.