set the node_offset parameter differently for each chunk
This was originally proposed by @zooko at
https://twitter.com/zooko/status/1060218512990363650

This isn't a security requirement per se, but the idea here is to
"disincentivize" certain implementation mistakes that might themselves
hurt security. For example, a dangerous optimization would be "check
each chunk to see if it's all null before hashing it, and if it is just
just hardcode the hash of the null chunk." That's dangerous because it
violates the constant time property; different inputs would hash at
different speeds depending on how many nulls they have, opening up side
channel attacks.

An optimization like that doesn't work for a regular serial hash
function, because each block's of input is hashed along with the entire
state so far, so the same block gives different output at different
positions. But it can work in a tree hash, since each chunk is designed
to be hashed indepdendently. Incrementing node_offset for each chunk
breaks this symmetry, "defeating" the optimization.

I was originally on the fence about this, in part because applications
that use sparse files might have a legitimate use case for optimizing
long runs of zeros. I also thought it was unlikely that anyone would
make this mistake in practice. But in my recent experiments, I've
learned that 1) the dangerous optimization above is very effective on
real data when most of your chunks are null, and 2) it has almost no
performance penalty when you're hashing random data. That makes it seem
much more realistic that someone might try this optimization. I've also
found that the node_offset mitigation also has no performance penalty,
and isn't too much trouble to implement in practice.