experiment with mmap'ing stdin on Unix
This is probably a bad idea on Windows, because I've seen unreliable
results from File::metadata after making an unsafe cast like this.

I probably won't ship this in any case, both because it's more unsafe
code that assumes thing std::fs::File doesn't guarantee, and because it
would be a foot-gun to have a pattern give good performance on Unix but
bad performance on Windows.