port Bao to BLAKE3
Things are coming full circle! The original Bao hash has evolved into
the BLAKE3 project. Now it's time to rework Bao itself in terms of the
upcoming BLAKE3 standard, focusing on the encoding and decoding
features. This branch will be published on 2020-01-09, together with the
rest of the BLAKE3 standard.

This commit replaces the internal tree hash implementation with calls to
the `blake3` crate (using a path dependency at commit
dc324a189e75f67d34464b9eab574f1851481c06, since it's not yet published
on crates.io). Unit tests are passing, but not the test vectors. The
next step is to update the Python implementation and re-generate the
vectors. After that there will be some substantial cleanup.

As part of this rework, I'm simplifying the code as much as possibe.
Previously, there were some tricky gymnastics involved in encoding and
decoding with as much parallelism as possible. Now that BLAKE3 is a
separate project with its own benchmarks, there is much less of a need
for Bao to be blazing fast. The decoding feature in particular will
almost always be used in a streaming setting, where the bottleneck is
either the network or the consuming application, rather than Bao itself.
For encoding, the bottleneck will usually be disk IO. The `bao hash`
command will just call into the `blake3` crate, so that will still be
very fast. (Potentially faster than before, in fact, since that crate
includes an AVX-512 implementation written in C.)